composer.json - Generalize version requirements
One of the issues complicating distribution via `composer` is that the
`composer.json` for the Civi-D8 module requires a specific version number.
This patch makes the `composer.json` more general -- for example, v5.1.1 of
`civicrm-drupal-8` would require v5.1.1 of `civirm-core`. This
reduces the amount of administration required for each release.