dev/drupal#107 - Fix user<=>contact sync used by installation workflow
Overview
--------

This fixes a bug which primarily manifests for the administration during a new Civi-D8 installation.

See: https://lab.civicrm.org/dev/drupal/issues/107

Before
------

Two ways to look at this:

(A) Main symptom - After enabling CiviCRM on D8, you navigate to the CiviCRM dashboard and get a weird error like:

> Sorry, due to an error, we are unable to fulfill your request at the moment. You may want to contact your administrator or service provider with more details about what action you were performing
> DB Constraint Violation - contact_id should possibly be marked as mandatory for DashboardContact,create API. If so, please raise a bug report.

(B) Formal description - If (for whatever reason) a D8 user is not linked to a Civi contact, then any attempt to work with the active user/contact will fail.

After
-----

(A) Main symptom - After enabling CiviCRM on D8, you can start using CiviCRM.

(B) Formal description - If (for whatever reason) a D8 user is not linked to a Civi contact, and if they try to use a `civicrm` page, then a
contact will be created and linked.

Technical Details
-----------------

This same technique has been active in Civi-D7 for years.

The code for this technique was present in Civi-D8, but it had some comments indicating that its significance wasn't understood at the
time of porting, and it was run at the wrong time (i.e. *after* the page-execution).